+++ Docs/articles/ReleaseNotes.md
@@ -11,6 +11,8 @@ The release versions are NuGet releases.
 * Added support for `Switch` element with "systemLanguage" selector attribute (see [#1176](https://github.com/svg-net/SVG/issues/1176)).
 * Enabled `IsTrimmable` and `IsAotCompatible` attributes on the assembly (see [#1184](https://github.com/svg-net/SVG/pull/1184)).
 * `SvgDocumentMetadata` will now keep existing children "as-is" when parsing. (see [#1219](https://github.com/svg-net/SVG/pull/1220)).
+* Added basic support for the `SvgStyle` element. (see [#1221](https://github.com/svg-net/SVG/issues/1221)).
+* Added support for `SvgCssVariablePaintServer` so declared CSS variables can be used as paint servers. (see [#1221](https://github.com/svg-net/SVG/issues/1221)).
 
 ### Fixes
 * Fixed `SvgGradientStop(SvgUnit, Color)` to apply the supplied color to `StopColor`. (see [PR #1217](https://github.com/svg-net/SVG/issues/1216))